+ Cleaning up code to eliminate warnings:   - Initialize all data members in constructor's initalizer list.   - Use static_cast to eliminate comparison of signed and unsigned integer types.
+ In Sides.cc, replace assignment with comparison operator...
  replace
	if (sideType[s] = this_side_type) {...}
  with
   	if (sideType[s] == this_side_type) {...}

+++ src/RTT_Format_Reader/SideDataIDs.cc
@@ -49,11 +49,12 @@ void SideDataIDs::readData(ifstream & meshfile)
     int dataIDNum;
     string dummyString;
 
-    for (unsigned i = 0; i < dims.get_nside_data(); ++i)
+    for (unsigned i = 0;
+         i < static_cast<unsigned int>(dims.get_nside_data()); ++i)
     {
 	Check(i<names.size() && i<units.size());
 	meshfile >> dataIDNum >> names[i] >> units[i];
-	Insist(dataIDNum == i+1,
+	Insist(static_cast<unsigned int>(dataIDNum) == i+1,
 	       "Invalid mesh file: side data ID out of order");
 	std::getline(meshfile, dummyString);
     }
