+++ src/RTT_Format_Reader/SideFlags.cc
@@ -49,10 +49,11 @@ void SideFlags::readFlagTypes(ifstream & meshfile)
     int flagTypeNum;
     string dummyString;
 
-    for (unsigned i = 0; i < dims.get_nside_flag_types(); ++i)
+    for (unsigned i = 0;
+         i < static_cast<unsigned int>(dims.get_nside_flag_types()); ++i)
     {
 	meshfile >> flagTypeNum >> dummyString;
-	Insist(flagTypeNum == i+1,
+	Insist(static_cast<unsigned int>(flagTypeNum) == i+1,
 	       "Invalid mesh file: side flag type out of order");
 	Check(i<flagTypes.size());
 	flagTypes[i] = new Flags(dims.get_nside_flags(i), dummyString);
